requirements.txt
clip @ git+https://github.com/openai/CLIP.git@04f4dc2ca1ed0acc9893bd1a3b526a7e02c4bb10
ftfy==6.0.3
grad-cam==1.3.7
librosa==0.8.1
ffmpeg-python==0.2.0
numpy>=1.19.2
opencv-python==4.5.4.60
pandas==1.3.4
Pillow==8.4.0
pytorch-lightning==1.1.3
regex==2021.10.8
scikit-learn==1.0
scikit-video==1.1.11
scipy==1.7.1
torch==1.11.0
torchaudio==0.11.0
torchvision>=0.8.2
tqdm==4.62.3
ttach==0.0.3
wav2clip==0.1.0
